+++ src/Trarizon.Library/TraTuple.cs
@@ -1,6 +1,0 @@
-﻿namespace Trarizon.Library;
-public static class TraTuple
-{
-    public static KeyValuePair<TKey, TValue> ToKeyValuePair<TKey, TValue>(this (TKey Key, TValue Value) tuple)
-        => new(tuple.Key, tuple.Value);
-}